# docs/static/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/t/{tenantCode}/interaction/{uid}/api/details": {
      "get": {
        "operationId": "InteractionController_getDetails",
        "parameters": [
          {
            "name": "tenantCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uid": {
                      "type": "string",
                      "example": "interaction-uid"
                    },
                    "prompt": {
                      "type": "string",
                      "example": "login"
                    },
                    "clientId": {
                      "type": "string",
                      "example": "web-app"
                    },
                    "missingScopes": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "email"
                      }
                    },
                    "mfaRequired": {
                      "type": "boolean",
                      "example": false
                    },
                    "idpList": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "provider": {
                            "type": "string",
                            "example": "google"
                          },
                          "name": {
                            "type": "string",
                            "example": "Google"
                          },
                          "protocol": {
                            "type": "string",
                            "example": "oauth2"
                          }
                        },
                        "required": ["provider", "name", "protocol"]
                      }
                    }
                  },
                  "required": [
                    "uid",
                    "prompt",
                    "clientId",
                    "missingScopes",
                    "mfaRequired",
                    "idpList"
                  ]
                }
              }
            }
          }
        },
        "summary": "Get interaction details",
        "tags": ["Interaction"]
      }
    },
    "/t/{tenantCode}/interaction/{uid}/api/login": {
      "post": {
        "operationId": "InteractionController_submitLogin",
        "parameters": [
          {
            "name": "tenantCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "mfaRequired": {
                      "type": "boolean",
                      "example": false
                    },
                    "passwordChangeRequired": {
                      "type": "boolean",
                      "example": false
                    },
                    "redirectTo": {
                      "type": "string",
                      "example": "/t/acme/interaction/uid"
                    },
                    "methods": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "totp"
                      }
                    },
                    "error": {
                      "type": "string",
                      "example": "invalid_credentials"
                    },
                    "retryAfterSec": {
                      "type": "integer",
                      "example": 60
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "summary": "Submit interaction login",
        "tags": ["Interaction"]
      }
    },
    "/t/{tenantCode}/interaction/{uid}/api/mfa": {
      "post": {
        "operationId": "InteractionController_submitMfa",
        "parameters": [
          {
            "name": "tenantCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionMfaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "mfaRequired": {
                      "type": "boolean",
                      "example": false
                    },
                    "passwordChangeRequired": {
                      "type": "boolean",
                      "example": false
                    },
                    "redirectTo": {
                      "type": "string",
                      "example": "/t/acme/interaction/uid"
                    },
                    "methods": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "totp"
                      }
                    },
                    "error": {
                      "type": "string",
                      "example": "invalid_credentials"
                    },
                    "retryAfterSec": {
                      "type": "integer",
                      "example": 60
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "summary": "Submit interaction MFA",
        "tags": ["Interaction"]
      }
    },
    "/t/{tenantCode}/interaction/{uid}/api/mfa/totp/enroll": {
      "post": {
        "operationId": "InteractionController_beginTotpEnrollment",
        "parameters": [
          {
            "name": "tenantCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "secret": {
                      "type": "string",
                      "example": "JBSWY3DPEHPK3PXP"
                    },
                    "otpauthUrl": {
                      "type": "string",
                      "example": "otpauth://totp/Auth:john?secret=JBSWY3DPEHPK3PXP"
                    }
                  },
                  "required": ["secret", "otpauthUrl"]
                }
              }
            }
          }
        },
        "summary": "Begin interaction TOTP enrollment",
        "tags": ["Interaction"]
      }
    },
    "/t/{tenantCode}/interaction/{uid}/api/mfa/totp/confirm": {
      "post": {
        "operationId": "InteractionController_confirmTotpEnrollment",
        "parameters": [
          {
            "name": "tenantCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionTotpConfirmationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "mfaRequired": {
                      "type": "boolean",
                      "example": false
                    },
                    "passwordChangeRequired": {
                      "type": "boolean",
                      "example": false
                    },
                    "redirectTo": {
                      "type": "string",
                      "example": "/t/acme/interaction/uid"
                    },
                    "methods": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "totp"
                      }
                    },
                    "error": {
                      "type": "string",
                      "example": "invalid_credentials"
                    },
                    "retryAfterSec": {
                      "type": "integer",
                      "example": 60
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "summary": "Confirm interaction TOTP enrollment",
        "tags": ["Interaction"]
      }
    },
    "/t/{tenantCode}/interaction/{uid}/api/password-change": {
      "post": {
        "operationId": "InteractionController_submitPasswordChange",
        "parameters": [
          {
            "name": "tenantCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionPasswordChangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "mfaRequired": {
                      "type": "boolean",
                      "example": false
                    },
                    "passwordChangeRequired": {
                      "type": "boolean",
                      "example": false
                    },
                    "redirectTo": {
                      "type": "string",
                      "example": "/t/acme/interaction/uid"
                    },
                    "methods": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "totp"
                      }
                    },
                    "error": {
                      "type": "string",
                      "example": "invalid_credentials"
                    },
                    "retryAfterSec": {
                      "type": "integer",
                      "example": 60
                    }
                  },
                  "required": []
                }
              }
            }
          }
        },
        "summary": "Submit required password change",
        "tags": ["Interaction"]
      }
    },
    "/t/{tenantCode}/interaction/{uid}/api/consent": {
      "post": {
        "operationId": "InteractionController_submitConsent",
        "parameters": [
          {
            "name": "tenantCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "redirectTo": {
                      "type": "string",
                      "example": "/t/acme/interaction/uid"
                    }
                  },
                  "required": ["redirectTo"]
                }
              }
            }
          }
        },
        "summary": "Submit interaction consent",
        "tags": ["Interaction"]
      }
    },
    "/t/{tenantCode}/interaction/{uid}/api/abort": {
      "get": {
        "operationId": "InteractionController_abortInteraction",
        "parameters": [
          {
            "name": "tenantCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "redirectTo": {
                      "type": "string",
                      "example": "/t/acme/interaction/uid"
                    }
                  },
                  "required": ["redirectTo"]
                }
              }
            }
          }
        },
        "summary": "Abort interaction",
        "tags": ["Interaction"]
      }
    },
    "/t/{tenantCode}/interaction/{uid}/api/mfa/webauthn-options": {
      "get": {
        "operationId": "InteractionController_getWebAuthnOptions",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "summary": "Get WebAuthn authentication options",
        "tags": ["Interaction"]
      }
    },
    "/t/{tenantCode}/interaction/{uid}/idp/{provider}": {
      "get": {
        "operationId": "InteractionController_redirectToIdp",
        "parameters": [
          {
            "name": "tenantCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect",
            "headers": {
              "Location": {
                "description": "Redirect target URL",
                "schema": {
                  "type": "string",
                  "example": "https://app.example.com/callback"
                }
              }
            }
          }
        },
        "summary": "Redirect to external identity provider",
        "tags": ["Interaction"]
      }
    },
    "/t/{tenantCode}/interaction/{uid}/idp/{provider}/callback": {
      "get": {
        "operationId": "InteractionController_idpCallback",
        "parameters": [
          {
            "name": "tenantCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect",
            "headers": {
              "Location": {
                "description": "Redirect target URL",
                "schema": {
                  "type": "string",
                  "example": "https://app.example.com/callback"
                }
              }
            }
          }
        },
        "summary": "Handle external identity provider callback",
        "tags": ["Interaction"]
      }
    },
    "/t/{tenantCode}/interaction/saml/{provider}/metadata": {
      "get": {
        "operationId": "InteractionController_samlMetadata",
        "parameters": [
          {
            "name": "tenantCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/samlmetadata+xml": {
                "schema": {
                  "type": "string",
                  "format": "xml"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "xml"
                }
              }
            }
          }
        },
        "summary": "Get SAML SP metadata",
        "tags": ["Interaction"]
      }
    },
    "/t/{tenantCode}/interaction/saml/{provider}/callback": {
      "post": {
        "operationId": "InteractionController_samlCallback",
        "parameters": [
          {
            "name": "tenantCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SamlCallbackDto"
              }
            }
          }
        },
        "responses": {
          "302": {
            "description": "Redirect",
            "headers": {
              "Location": {
                "description": "Redirect target URL",
                "schema": {
                  "type": "string",
                  "example": "https://app.example.com/callback"
                }
              }
            }
          }
        },
        "summary": "Handle SAML identity provider callback",
        "tags": ["Interaction"]
      }
    },
...
  },
  "info": {
    "title": "Auth API",
    "description": "OIDC Authorization Server API Reference",
    "version": "0.1.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Service origin"
    },
    {
      "url": "/api",
      "description": "Vite proxy"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session"
      }
    },
    "schemas": {
      "InteractionLoginDto": {
        "type": "object",
        "properties": {}
      },
      "InteractionMfaDto": {
        "type": "object",
        "properties": {}
      },
      "InteractionTotpConfirmationDto": {
        "type": "object",
        "properties": {}
      },
      "InteractionPasswordChangeDto": {
        "type": "object",
        "properties": {}
      },
      "SamlCallbackDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}